+++ maildir_deduplicate/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2010-2016 Kevin Deldycke <kevin@deldycke.com>
+# Copyright (C) 2010-2017 Kevin Deldycke <kevin@deldycke.com>
 #                         and contributors.
 # All Rights Reserved.
 #
@@ -24,7 +24,7 @@ import logging
 import sys
 
 
-__version__ = '2.0.2'
+__version__ = '2.1.0'
 
 
 PY2 = sys.version_info[0] == 2